Move column encoder 1 into separate file
